final final offline

+++ Offline-1/1805086.sh
@@ -13,7 +13,8 @@ files=$(ls Submissions)         # list the files in the directory submissions
 
 
 for file_1 in $files; do        # for each file in the directory
-    if [ -d Submissions/$file_1 ]; then         # if the file is a directory proceed
+    if [ -d Submissions/$file_1 ] && [ -f Submissions/$file_1/$file_1.sh ]; then         # if the file is a directory proceed
+        echo $file_1
         chmod 554 Submissions/$file_1;                  # give the permission to execute permission to owner and group 
                                                         # and readpermission to all (folder)
         chmod 554 Submissions/$file_1/$file_1.sh;       # same permission to sh files....
@@ -38,27 +39,30 @@ for file_1 in $files; do        # for each file in the directory
         if [ $score -lt 0 ]; then           # if the score is less than 0
             score=0;                        # set the score to 0
         fi
+
+        #now copy checker inmlementation
+        #compare the current student's file with all the other students' files
+        #if there is a match, then the student gets -score
+        for file_2 in $files; do
+            if [ -d Submissions/$file_2 ] && [ -f Submissions/$file_2/$file_2.sh ]; then            # if the file is a directory and there is a sh file in the directory
+                if [ "$file_1" != "$file_2" ]; then                                                 # if the file is not the current student's file
+
+                    # compare the current student's file with the other student's file
+                    diff Submissions/$file_1/$file_1.sh Submissions/$file_2/$file_2.sh --ignore-all-space > Submissions/$file_1-$file_2-diff.txt;
+                    
+                    # count the number of lines in the diff file(this is the number of differences)
+                    val=$(grep -c "^[<>]" Submissions/$file_1-$file_2-diff.txt);
+                    if [ $val -eq 0 ]; then             # if the number of differences is 0
+                        score=$((-1 * $score));         # score is set to -score
+                    fi
+                    rm Submissions/$file_1-$file_2-diff.txt;            # remove the diff file
+                fi
+            fi
+        done
+    else
+        score=0;            # if the file is not a directory or there is no sh file in the directory, set the score to 0
     fi
 
-    #now copy checker inmlementation
-    #compare the current student's file with all the other students' files
-    #if there is a match, then the student gets -score
-    for file_2 in $files; do
-        if [ -d Submissions/$file_2 ]; then             # if the file is a directory
-            if [ "$file_1" != "$file_2" ]; then         # if the file is not the current student's file
-
-                # compare the current student's file with the other student's file
-                diff Submissions/$file_1/$file_1.sh Submissions/$file_2/$file_2.sh --ignore-all-space > Submissions/$file_1-$file_2-diff.txt;
-                
-                # count the number of lines in the diff file(this is the number of differences)
-                val=$(grep -c "^[<>]" Submissions/$file_1-$file_2-diff.txt);
-                if [ $val -eq 0 ]; then             # if the number of differences is 0
-                    score=$((-1 * $score));         # score is set to -score
-                fi
-                rm Submissions/$file_1-$file_2-diff.txt;            # remove the diff file
-            fi
-        fi
-    done
     echo "$file_1, $score" >> output.csv;            # echo students_id, score in a file
 
 done
